+++ chat-components/src/components/notificationpane/interfaces/common/INotificationPaneClassNames.ts
@@ -1,0 +1,11 @@
+/**
+ * This interface will have the common class names properties and is inherited by each scenarios.
+ */
+export interface INotificationPaneClassNames {
+    containerClassName?: string;
+    titleClassName?: string;
+    subtitleClassName?: string;
+    dismissButtonClassName?: string;
+    hyperlinkClassName?: string;
+    notificationIconClassName?: string;
+}